Move copied classes to dedicated package
Use org.wikimedia.gobblin.copy package for copied classes instead of
their original gobblin package. This should prevent errors are runtime
possibly due to classpath equality.
Also adds comments in copied classes for clarity.
