+++ gobblin-wmf-core/src/test/java/org/wikimedia/gobblin/kafka/TestUtils.java
@@ -22,9 +22,9 @@ import com.google.common.base.Optional;
 import org.apache.gobblin.configuration.ConfigurationKeys;
 import org.apache.gobblin.configuration.State;
 import org.apache.gobblin.configuration.WorkUnitState;
-import org.apache.gobblin.kafka.client.Kafka1ConsumerClient;
+import org.wikimedia.gobblin.copy.Kafka1ConsumerClient;
 import org.apache.gobblin.source.extractor.WatermarkInterval;
-import org.apache.gobblin.source.extractor.extract.kafka.KafkaSource;
+import org.wikimedia.gobblin.copy.KafkaSource;
 import org.apache.gobblin.source.extractor.extract.kafka.MultiLongWatermark;
 import org.apache.gobblin.source.workunit.WorkUnit;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
